refactor: Extract UI functionality into dedicated module
- Move all UI helper functions from main.rs to src/ui/interactive.rs
- Extract run_interactive_ui implementation to UI module
- Update main.rs to use UI module functions via clean imports
- Remove duplicate code and unused imports
- Maintain identical application behavior and functionality
- All tests pass, clippy clean, code properly formatted

This refactoring improves code organization by separating UI concerns
from main application logic, making the codebase more maintainable
and following better separation of concerns principles.
